[update] 再帰処理を廃止し、コメントを同じ階層で扱うように変更
コメントに対するコメントを同じ階層で扱うために、再帰処理を廃止し、ParentComment.vueとChildComment.vueを作成しました。それに伴い、不要になったCommentList.vueとCommentItem.vueを削除しました。現在、子コメントに対する返信フォームが表示されない問題があります。
